Remove inheritance from pacemaker::stonith of pacemaker::corosync
When using the pacemaker module together with an ENC, (Satellite 6)
we ran into the problem that we want to include both the pacemaker::stonith
and the pacemaker::corosync classes into our manifest in order to assign parameters
to both. Here we ran into a problem since the pacemaker::corosync class ended up redefined.
We have here removed the inheritance and instead put a require in the pacemaker::stonith class