Add $verify_on_create ability for all resources.
Implements a workaround for the rare condidtion where "pcs resource
create" may succeed on the command line, but pacemaker is not updated
with the new resource.

When $verify_on_create is true, make sure the resource exists after
$try_sleep with an additional "pcs" call.  Retry up to $tries per the
usual behaviour.

When $verify_on_create is false (the default), behaviour is exactly
the same as before, i.e. only the return value from "pcs resource
create" is checked to determine whether the resource creation was
successful or not.