Add support for Redundant Ring Protocol (RRP)
Corosync's purpose is to keep a cluster of nodes running.
If a node loose connectivity it can be fenced and pulled out the
cluster.

Currently this modules allow to setup corosync using only a single ring
(ie. a single network), ring0. That means that if for any reason the network
falls entirely the cluster will be down.

By enabling a second ring (ie. a second network), ring1 we gain
fault-tolerancy on the whole network. So if we have ring0 and ring1 and
two totally different network, if ring0 falls, cluster will remain
operational and all the traffic will go through ring1. This is the
default behavior with rrp_mode passive).

We add a new parameter `$cluster_members_rrp` that defines the nodes and
their interface they are in. `pcs cluster setup --start` will run with
this parameter if it is specified else default to `$cluster_members`.

Both parameters are necessary because `$cluster_members_rrp` is
incompatible with the `pcs cluster auth` command.

An example setup can be

```
class { '::pacemaker::corosync':
   cluster_members => 'node01.example.com node02.example.com node03.example.com',
   cluster_members_rrp => 'node01.example.com,node01.ring.example.com node02.example.com,node02.ring.example.com node03.example.com,node03.ring.example.com',
   setup_cluster   => true,
  }
```

The outputs of `pcs cluster corosync` looks like the following :

```
[root@localhost ~]# pcs cluster corosync
totem {
version: 2
secauth: off
cluster_name: clustername
transport: udpu
rrp_mode: passive
}
nodelist {
  node {
        ring0_addr: node01.example.com
        ring1_addr: node01.ring.example.com
        nodeid: 1
       }
  node {
        ring0_addr: node02.example.com
        ring1_addr: node02.ring.example.com
        nodeid: 2
       }
  node {
        ring0_addr: node03.example.com
        ring1_addr: node03.ring.example.com
        nodeid: 3
       }
}
quorum {
provider: corosync_votequorum
}
logging {
to_syslog: yes
}
```