Fix invalid boolean check for properties' values
The property resource checks that if 'ensure' is set to present, then a
value must be set; This check is wrong because false can actually be a
valid value for the property. In reality, what we want to check for is
if the value has been set at all. So this change introduces a check for
undef.