Fixes for lint compliance and working specs.
This fixes agent generator to have the generated files passing lint
test.  It also adds a rake task.

The stonith/ files were generated using it.  fence_xvm is documented to
be updated manually in the variable.sh file.  It will need to be
investigated, and if possible refactored to be able to work with the
generator for simpler updates.

Co-Authored-By: Jason Guiditta <jguiditt@redhat.com>
Change-Id: I246eee81de6e8148f4f8c01e6deb42b20775079a