Colocation constraint always has master role.
All colocation contraint are created with master role.  This is because
currently the code test the if @resource[:master_role] is true.  As the
parameter return the symbol :false, this is considered as a true value.

This fix the problem by making sure the parameter is cast into a
TrueClass, FalseClass.

Closes-Bug: 1583792

Change-Id: Ic00e49a71130b7f00f33a5f698b2de6e5608dcce