CI: specinfra, xenial, lint, puppet bug.
Fix beaker job for pacemaker.  When
https://tickets.puppetlabs.com/browse/BKR-851 is fixed, remove it.

The xenial node definition is now gated for Newton, add it to the mix.

Fix puppet puppet-lint-absolute_classname-check.

Work around https://github.com/puppetlabs/puppet/pull/2633 where contain
cannot have a absolute class name as parameter in puppet 3.6.

For Ubuntu Xenial we are bitten by this
https://tickets.puppetlabs.com/browse/BKR-821, where puppet is installed
in /opt/puppetlabs/bin

The package pacemaker-mgmt is not available on Ubuntu Xenial, adjust to
this.  Dbus is required by pacemaker and not installed as a dependency:
http://logs.openstack.org/36/332836/7/check/gate-puppet-pacemaker-puppet-beaker-rspec-ubuntu-xenial/5e1ad11/console.html#_2016-06-23_16_03_00_952026

Add necessary support files for reno.

Adjust the metadata.json.

Change-Id: I22e207f758b94586ebf2333668c092fff3aaebac