Pin beaker to < 3.0.0
Due to the requirement of ruby 2.2.x for beaker 3.1.0, we need to pin to
a working version. Version 3.0.0 of beaker had a bad reference that lead
to an older version of serverspec being pulled in. So we need to pin to
the last known good version which was in the 2.x series.

Change-Id: I0813ee882b1ecefa4eeea342c1594b53360574ab