Add minimal location rule support to resource creation
Currently it is impossible to create a location constraint and a
resource in a non racy way. For example it is not currently possible
to do something like the following:
pacemaker::resource::ocf { 'galera-master': ... }
pacemaker::constraint::location_rule { 'location-galera':
  resource           => 'galera-master',
  expression         => 'galera-role eq true',
  resource_discovery => 'exclusive',
}

The reason the above does not work is that after the resource creation
the pacemaker is free to start the resource wherever it wants and this
can break certain deployments. There are two blessed ways to solve this
via pcs:
1) pcs resource create galera-master ... --disabled
   pcs constraint location rule ...
   pcs resource enable galera-master
2) pcs cluster cib > cib.xml
   pcs -f cib.xml resource create galera-master ...
   pcs -f cib.xml constraint location rule ...
   pcs cluster cib-push cib.xml --config

Both approaches have been tried and the first one, with the addition of
a location_rule parameter to all resources seems to be the less
intrusive one (does not break current users and requires little effort
to use). The second approach was prototyped here:
https://review.openstack.org/402219

Change-Id: I8d78cc1b14f0e18e034b979a826bf3cdb0878bae