Relax requires when creating resources on remote nodes
In the composable HA architecture we must be able to create
properties and resources from a pacemaker remote node (which
does not have neither corosync nor pacemaker running).
In order to be able to do that we need to relax the requires
and remove the "Exec['wait-for-settle']" which only exists
in the corosync class, which does not exist on a pacemaker remote
node.

The onus is on the user to make sure that the cluster is functional
before trying to create a resource from a pacemaker remote node.
In TripleO we do this by creating the resources only on step2.

The pacemaker_remote configuration for the remote node itself was
never really designed in to the current module, so tweaking the
existing corosync class to cater for remote nodes would mean rewriting
most of the class and is not worth it when we are actually trying
to merge with the "new" puppet pacemaker module in the future.

Currently the pacemaker_remote node configuration happens in a tripleo
profile: I92953afcc7d536d387381f08164cae8b52f41605

Note: Creating resources on a pacemaker remote node only works
with pcs-0.9.152 or later (CentOS/RHEL 7.3 and later)

Change-Id: I5379b804757d23cae694943b54e7d9256f3ca669