Move cluster properties implementation to ruby
Currently cluster properties are simply implemented with a few
pcs Exec calls. Reimplement this functionality in ruby so that
we can leverage the newly introduced common pcs() functions.

This move also makes the code a bit simpler as it is cleaner to
write the code to deal with the following pcs bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1302010.

Change-Id: I9cff31b903e9f2ff6d1a81da99cabea6a7c8b639