Do not operate on the live CIB with pcs commands
When doing composable HA we typically create different pacemaker
resources on the nodes that require it (i.e. not from a single node).

This is problematic due to https://bugzilla.redhat.com/show_bug.cgi?id=1404233
The problem is that when node A creates resource foo and at the same
time node B creates resource bar, there is an potential race condition
because if both node A and B get a copy of the CIB at the same time,
the latter one to push the cib back (via pcs) will win and overwrite
the changes of the former one.

The most solid approach we discussed with the pcs team is twofold:
Use pacemaker's diff feature so that only a specific diff is sent on the
wire and not the whole CIB. This will make sure we avoid this race
because we never create overlapping resources from the same nodes. This
differential change requires that we do the following:
- Snapshot the CIB via "pcs cluster cib snapshot.xml"
- Make a copy of the CIB "cp snapshot.xml snapshot.xml.org"
- Make all changes via pcs -f snapshot.xml ....
- Push back the changes via an upcoming pcs command:
  pcs cluster cib-push snapshot.xml --diff=snapshot.xml.orig

This is all handled transparently by the backup_cib() and push_cib()
functions introduced earlier.

Change-Id: Id694585bb90c49f17e7a43f3ae555c6bd94945cb