Do not apply the cib via --config when pcs does not support diffing
In change I47558b5133b1404226b67b54eeeb398411ab0c53 we added the
'--config' when pushing the CIB. It turns out there is a case, when pcs
version is previous to pcs-0.9.152-10.el7_3.1.x86_64 where this is
detrimental and will actually *not* return an error if another node
pushed the CIB while another node dumped it and pushed it in the
meantime. This means that some nodes would remove the properties of
other nodes, which could get us with stuck pacemaker resources. One
phenomenon of this has been observed here:
http://logs.openstack.org/89/362389/26/check-tripleo/gate-tripleo-ci-centos-7-ovb-ha/a127a87/logs/

Note that this change is a noop once pcs-0.9.152-10.el7_3.1.x86_64
is shipped (should happen this week).

Change-Id: Ib931adaff43dbc16220a90fb509845178d696402