Add support for ipv6_addrlabel with IPaddr2 RA
So currently when we create an IPv6 VIP in puppet the
services running on the node where the VIP is, are susceptible
to using the VIP as a source address in their communication.
When the VIP relocates to another node, those TCP connections
using the VIP as a source will be hanging until a TCP timeout
kicks in.

There are many ways to guide the kernel in a source address selection
with IPv6 (call "ip -6 addr add <...> " with the "noprefixroute", use
address labels, or simply set the deprecated option on the created
ip address). We are choosing address labels because it is already
supported in the resource-agents package and seems the recommended
way of doing things.

Partial-Bug: #1686357

Change-Id: Ifd42683efae2cb29be3fcc1d5afdb62dad1bd82d
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Co-Authored-By: John Eckersberg <jeckersb@redhat.com>