Fix up race when setting a property
Back in January when we moved the property support implementation to
ruby the 'require' line got lost. This makes it so that there is a race
where we set a property but the cluster might not be fully up and so we
might fail when dumping the CIB:

May  4 13:46:44 centos-7-2-node-osic-cloud1-s3500-8719947-570244
os-collect-config:  Error: unable to get cib
Error: /Stage[main]/Pacemaker::Stonith/Pacemaker::Property[Disable
STONITH]/Pcmk_property[property--stonith-enabled]: Could not evaluate:
backup_cib: Running: /usr/sbin/pcs cluster cib
/var/lib/pacemaker/cib/puppet-cib-backup20170504-29225-fetp7k failed
with code: 1 ->

Am not entirely sure why I saw this only today in a failed CI job, but
it is likely that being a race it won't show up all too often and likely
only when we force the stonith property to be false by default.

Change-Id: Iaebc9b93182f35a4cf143599bd777a4ed5f2de17
Closes-Bug: #1688322