Add retries for pcs cluster setup
In bug #1677312 we encountered an error where pcs cluster start --all
would return 1, due to pcsd on one node being in the middle of a restart
(due to ssl keys exchange).

It turns out this can also happen during the pcs cluster setup call, so
let's add the proper retries there as well. This has
been spotted a couple of time in largish clusters.

Closes-Bug: #1692923

Change-Id: Ia5d820c97d5dfdaaa03aff2721612937e0efa2e6