Fix brokenness when changing the value of an existing property
In TripleO we set the 'stonith-enabled' property to false
at the beginning of the installation and then set it to true
towards the end of the install. The reason this is not working
is that the exist? function only greps for the property's
existance and does not verify if the value is the one
that the puppet module is trying to set.

Fix this by making sure we also check for the expected value.
We do this only if the expected 'ensure' status of the resource
is set to 'present'

Change-Id: Ia28d14d1763373494ae0bfc37d53d1fabc304d1e
Closes-Bug: #1717531