Add pcmk_stonith provider
This new pcmk_stonith provider exists in order to create
stonith resources and location constraints using the common
pcs() function. The reason for this is that this function
contains the retry logic needed in case the CIB gets updated
from another node.

In this provider we also create the location constraint in
a race-free way by first creating the stonith resource in
disabled mode, then we create the location constraint and
finally we enable the stonith resource.

This is to avoid races when creating stonith resources from
multiple nodes which is the how they are created within
TripleO.

Change-Id: I424302bbf8d0d5f233e3a7debc082be1c9a170bb
Partial-Bug: #1717566