Allow empty pcmk_host_list
While this is not needed for the currently automatically-generated
fence agents, in the upcoming fence_compute agent we actually want to
allow a user to be able to not set the 'pcmk_host_list' parameter when
calling pcs. Currently when it is set to 'undef' it will be replaced
with the output of $(crm_node -n). Let's add an explicit empty value
for the cases where we do *not* want to use this parameter at all.

Also make sure we do not create the constraint when pcmk_host_list is
empty.

Change-Id: I76c7ff97e5c2bba42f5ceb507686250493efa127