Generate fence_ironic automatically
When I315f9bb78a1296f691dadaeb39fb0a48df1d5f06 got merged the xml files
for fence_ironic were added but the actual manifest that was merged into
manifests/stonith/fence_ironic was manually generated.

That means that there are a bunch of parameters that do not work in the
current release (and probably never have?). For example:
- fence_ironic takes 'password' as an option:
  https://github.com/ClusterLabs/fence-agents/blob/master/fence/agents/ironic/fence_ironic.py#L53).
  But in the manifest we have 'passwd':
  https://github.com/openstack/puppet-pacemaker/blob/master/manifests/stonith/fence_ironic.pp#L85

Since this one and a few other options could never really have worked,
let's just autogenerate the file from xml properly. The xml has the
last word on which parameters work in any case.

Additionally there never has been a released fence-agents-ironic
package, so the package relationship cannot have worked unless the user
created a package himself. In order to fix that we allow a special
package name 'None' to be passed to the autogenerator which will skip
generating the package dependencies.  In this case it is the user's duty
to have the binary on the system.  We can change this back once a
fence-agents-ironic package does get released.

Closes-Bug: #1719389

Change-Id: Ie9e61c44fd5dd186f521f98ef2fd637005b00ee9