Add meta attribute support to stonith manifests
Stonith devices can also take a 'meta' attributes that
describes further characteristics of their behaviour.
We add a 'meta_attr' parameter that can be used to specify
additional meta parameters when invoking pcs.

Note: fence_xvm was done by hand as it is not autogenerated.

Tested with the upcoming fence_compute manifest (autogenerated)
and meta attributes are correctly set when meta_attr is not
undef:
$ pcs stonith show stonith-fence_compute-fence-nova |grep Meta
  Meta Attrs: provides=unfencing

Change-Id: I3648b358ca9799c2a430878ca5b17f570914a06d