Move location_rule build to a common function
Both in a bundle and in a resource we currently duplicate a bunch of
code in order to prepare the string to pass to pcs to create the
location rule constraint for resources. Let's split out the code to
create the string in a common function under pcmk_common.

Let's then leverage it from both bundles and resources.

Change-Id: I660cead06b5773aab473da1874e1edf5fb03e95a