Detect when a bundle or a resource needs updating
With this change we actually make it so that if we detect that a
resource or a bundle have changed (as compared to the existing already
running resource/bundle) we actually update the cluster with the new
resource. This does happen *only* when the "deep_compare" parameter is
set to true.

Note that when pacemaker updates a resource it will automatically
restart it. This means that if puppet gets reapplied it restart the
resource that changed.

This has been tested in the following environment (applying all the patches in this
series):
A three-node cluster with the following resources:
A) Docker bundle
B) An ocf resource (rabbitmq cloned) with a location rule constraint
C) A VIP resource

The following tests were made
1) Successful deployment of the cluster
2) Confirmed that a rerun of the unchanged puppet manifest triggers no
   restarts (no deep_compare hiera keys are set to true)
3) Set the following three hiera keys and observe that the unchanged
   puppet manifest triggers no restarts/changes:
     pacemaker::resource::ip::deep_compare: true
     pacemaker::resource::bundle::deep_compare: true
     pacemaker::resource::ocf::deep_compare: true
   Also verify that indeed the 'deep_compare' branch of the code is
   executed
4) Change the IP resource definition (change the netmask), make sure
   'pacemaker::resource::ip::deep_compare' is set to true, and observe
   that the VIP gets changed and restarted:
   Notice: /Stage[main]/Main/Pacemaker::Resource::Ip[my_ip]/Pcmk_resource[ip-172.16.11.97]/ensure: created
5) Rerun step 4) and observe that no changes took place
6) Change the OCF resource definition (change a rabbitmq parameter),
   make sure 'pacemaker::resource::bundle::deep_compare' is true and
   observe that the rabbitmq-clone resource gets restarted:
   Notice: /Stage[main]/Main/Pacemaker::Resource::Ocf[rabbitmq]/Pcmk_resource[rabbitmq]/ensure: created
7) Rerun 6) and observe that no changes took place
8) Change the bundle definition (lower the replicas number from 3 to 2),
   makre sure 'pacemaker::resource::bundle::deep_compare' is set to
   true, and observe that the bundle gets restarted:
   Notice: /Stage[main]/Main/Pacemaker::Resource::Bundle[test_bundle]/Pcmk_bundle[test_bundle]/ensure: created
9) Rerun 8) and observe that no changes took place

Change-Id: I1730e8e2b3d52690d4e006e588b4819662d6b2a1