Move the tests around so pcmk_resource_has_changed? does not get triggered
There is likely a bug lurking in how we do deep_compare of
pacemaker_remote resources. This should not even have been
triggered unless the deep_compare parameter is set to true.
Let's move the tests around so, as per design, we do not
even invoke the functions to compare resources in depth.

Change-Id: I5e7c28794db8b8b60d9d7598a1bb9f8b758484e0