Use crm_diff for pcmk resource deep_compare
Detect if crm_diff is fixed and if it is favour the crm_diff resource
comparison. Also introduce a 'update_settle_secs' parameter
for resources that can be used to override the amount of time (s)
to wait for the cluster to settle after updating a resource.

Tested this as follows:
1) Normal deploys work fine (the code never gets triggered anyway)
2) Redeploy with no changes -> pcmk resources are untouched
   (the code does not get triggered anyway)
3) Redeploy by adding the following hiera keys:
   ExtraConfig:
     pacemaker::resource::bundle::deep_compare: true
     pacemaker::resource::ip::deep_compare: true
     pacemaker::resource::ocf::deep_compare: true
   -> pcmk resources are untouched
4) Redeploy with keys at 3) but with changes that trigger a restart
   -> pcmk resources are restarted and cluster waits for the
      restart before continuing

Change-Id: I62d5662327333e17e1c32a029695b3d3a904ca10