Fix rspec test on ubuntu os.
There seems to be an issue with rspec-puppet where it doesn't have the
necessary facts for getting the right initsystem on debian's like
os[1].  It may then happen that it detects a runit system and
obviously fails on service checks.

We force that fact to systemd as this is true for all supported os.

[1] https://github.com/rodjek/rspec-puppet/issues/629

Change-Id: I8b406c33ae8467d54f1b8502fbde5dfd1fbb1204
Closes-Bug: #1783302