Deep_compare fix with location constraints and better debugging
This fixes one use case where a version of pcs might not remove the
associated location constraint from the offline CIB, so when
we tried to readd it would fail. We simply add --force when readding
it as the end result is still the same and the CIB won't change in
that regard anyway.

While we're at it we make the debug messages a bit clearer to ease
debugging.

Tested by Marian Krcmaric <mkrcmari@redhat.com> as well as by me.

Change-Id: I409812f1ec72ed450edcbdfd9849164b147eae8f