Rely on path for CLI calls when possible
We currently hard-code /usr/sbin/{pcs,crm_simulate,crm_resource} in a
few places. This is undesirable as it makes things difficult to
override, especially in the case where in a container you want to bind
mount those utils from the host in a specific-path.

Let's just call the utilities without the full path when the PATH
environment variable is set. It is the polite thing to do in case
an operator wants to use different versions of the tools in a different
path.

Tested by deploying tripleo with containers where pcs was both
bind-mounted in a specific location and with the standard /usr/sbin/pcs.

Related-Bug: #1771612

Change-Id: I933f101c82b9abad330fab622733e40cf15ef49b