Fix reconnect_interval parameter
There is a typo in the logic around reconnect_intervl
which makes it skip setting the interval.

Let's fix that by making the code more explicit.
Tested with a custom reconnect interval set via:
ExtraConfig:
  pacemaker_remote_reconnect_interval: 120

And obtained the correct state of the remote resource:
[root@controller-0 ~]# pcs resource show compute-0 |grep reconnect
  Attributes: reconnect_interval=120 server=172.17.1.14

Before the fix there would be no reconnect_interval
attribute.

Change-Id: I5af578b5b838deec7beb2d5a9bb83fa6bb6293b7
Related-Bug: #1797041