Skip some tests for Ubuntu 18.04 and puppet 4.8.
We already had an issue and we fixed it with an hack[1]. Puppet would
assign runit service provider for Ubuntu.  runit being the fallback
when nothing else match.

As a matter of fact puppet doesn't properly detect Ubuntu 18.04
systemd usage before puppet 5.  For instance here the code for the 4.x
branch[2]:

    defaultfor :operatingsystem => :ubuntu, :operatingsystemmajrelease => ["15.04","15.10","16.04","16.10","17.04","17.10"]

So let's get rid of those tests for 4.8 and Ubuntu 18.04.

I only test < 5 as the first maintained branch is 5.3 and has the
support for Ubuntu 18.04 services.

[1] https://bugs.launchpad.net/puppet-pacemaker/+bug/1783302
[2] https://github.com/puppetlabs/puppet/blob/4.10.x/lib/puppet/provider/service/systemd.rb#L28

Change-Id: I9828b9a837c652fdc87131a84c73591c4e84d387
Closes-Bug: #1802140