Match node properties more strictly
We currently do the following in order to detect if a property
is already set with the correct values:

  cmd = "property show | grep #{property}"
  if not_empty_string(node)
    cmd += " | grep #{node}"
  end
  if not_empty_string(value)
    cmd += " | grep #{value}"
  end
  cmd += " > /dev/null 2>&1"
  ret = pcs('show', @resource[:property], cmd, @resource[:tries], @resource[:try_sleep])
  Puppet.debug("property exists: #{cmd} -> #{ret}")
  return ret == false ? false : true

The problem above is that the above will match in the presence of the following
lines:
controller-1: cinder-volume-role=true galera-role=true haproxy-role=true rabbitmq-role=true redis-role=true rmq-node-attr-last-known-rabbitmq=rabbit@controller-1
controller-11: cinder-volume-role=true galera-role=true haproxy-role=true rabbitmq-role=true redis-role=true rmq-node-attr-last-known-rabbitmq=rabbit@controller-1

Let's make sure the grep for the node includes the terminating ':' so as to match
the exact node-name and not trip up on a subset. Let's also make sure
that the match is exactly on the key=value match since there can be many
properties on a single line.

Tested by deploying an overcloud successfully, changing a property and
then redeploying to verify that the property has been set back to the
correct value.

Change-Id: I5d3353d221bd33cbb0f9393c10721f9d057f2c8b
Closes-Bug: #1805786