Initial pcs 0.10 support
pcs 0.10 has changed a few things around cluster authentication
so we need to parametrize things depending on the pcs version on the
system.

We also make sure that if you use the cluster_members_addr parameter
it stays compatible with pcs 0.9 by only returning the node names
and not the nodenames + addr=... strings.

Tested this on a standalone env with pcs 0.10 (rhel8os + f28 containers
no pre-installed pcs rpm) and on an OSP14 using pcs 0.9

Change-Id: I8693cc0fdd51be155273e0dba8bc14b3db79ba8d