Add cluster_members_addr parameter to specify ip addresses used in cluster setup
Due to https://bugzilla.redhat.com/show_bug.cgi?id=1659389 we need to be
able to specify the exact ip addresses being used in the cluster setup.
This support exists also for corosync 2.x based systems, but seems to be
quite a necessity when using corosync 3.x + knet. So let's add this
feature in order to cover for the next Centos/RHEL release.

Change-Id: I03d23177c7a913127ffc2f6d01a6318cb80d09c5