Introduce a pcs_without_push function
With the new pcs 0.10.x version we cannot create pacemaker remotes like
we did previously.(pcs resource create foo ocf:remote ...). It will need
to run a bunch of pcs commands that cannot work on an offline CIB (i.e.
we cannot use the pcs() function which does the usual cib dump/push
cycle). So let's introduce a pcs_without_push() function which invokes
pcs without doing a cib push/dump mechanism.

The commands intended for this function are synchronous in the cluster
in any case.

Change-Id: Ia162322d969ac4909a187af446d1316aa788d94d