Scale up logic with extra IP addresses for pcs 0.10
Nowadays we can provide IP addresses when adding a host into a
pacemaker cluster. The pcmk_cluster_setup function for initial setup
of the cluster has this already implemented, but it was still missing
from the pcmk_nodes_added function and the scale-up logic. Support for
scaling up with pcs 0.10 and the explicitly provided IP addresses is
being added in this commit.

Change-Id: If5baef83c739c7ccdbe346c75401f232ab092d20
Closes-Bug: #1835217
Related-Bug: #1831690