pcs 0.10: authenticate nodes before adding them to the cluster
With pcs 0.10 and RHEL/CentOS 8, every node that needs to be
added into a running cluster muste be authenticated first in pcsd.

So when replacing a cluster node or automatically scaling up the
cluster, make sure that we run a "pcs host auth" command prior to
any "pcs cluster node add".

Change-Id: I88c1188b2202e0cf70b82db629c3dd748ad01bf3
Closes-Bug: #1839209