Make pacemaker_resource serializable
Change the internal storage of resource operations from a Set to an
Array of Hashes. This prevents an error when a Puppet Agent tries to
serialize the data to its transaction store and fails to re-read it
with:

Error: Transaction store file
/opt/puppetlabs/puppet/cache/state/transactionstore.yaml is corrupt
((/opt/puppetlabs/puppet/cache/state/transactionstore.yaml): Tried to
load unspecified class: Set); replacing

To support the change, the handling of operations is generally stricter,
preferring to raise an error on unexpected input instead of silently
dropping invalid entries or attempting to convert them.

Change-Id: I4bee307c81d606981574cd0004150f6f3f31f81e