Try and reauthenticate if pcsd status $::hostname fails
If we fail the local authentication via pcs, let's try and reauthenticate
This might happen if /var/lib/pcsd/tokens got corrupt or if we upgraded pcs versions
and authentication is not working even though the hacluster user has not changed

Tested a deploy and a redeploy and it all worked correctly. This should
give us extra automatic reauth in case something goes wrong.

Change-Id: Icaed03fe167ca08be60ca9f6851c8fc7b64411d0