Add support for promoted-max meta attribute
Master/Slave bundles used to have an attribute "masters" to
specify how many replicas can be promoted to Master state. That
attribute has been deprecated and renamed "promoted-max".

Expose the new attribute in the puppet resource bundle, and use
it automatically if an instantiated resource is configured with
deprecated attribute "masters" with pacemaker 2.

Closes-Bug: #1884108
Change-Id: I3033563ba8a28298d63db42dfa2df1960b685e45