Fix up upgrade coordination around authkey
Preamble: RHEL/CentOS 8 use pcs 0.10 which manages
/etc/pacemaker/authkey for all remotes and core cluster nodes.
That was not the case with pcs 0.9 where we had to manage that file
on our own.

This patch does three things:
1) It introduces a force_authkey parameter so we can use the existing
   authkey file even when pcs 0.10 is being used
2) It introduces a parameter called force_oldstyle to forcibly create remotes in
   the pcs 0.9 way (which is different than the pcs 0.10 way)
3) On remotes it makes sure that any resource is ordered *after* the
   pcsd/pacemaker_remote service

1) and 2) are needed to be able to run puppet when the core cluster is
on RHEL/CentOS 8 whereas the remotes are on RHEL/CentOS 7. In this
situation we need to keep using the authkey we manage on both remotes
and core cluster nodes *and* we need to be creating the remotes in the
old pcs 0.9 way.

Tested this with multiple IHA FFU runs and also on a plain non-IHA Train
deployment with this change.

Related-Bug: #1888398

Change-Id: I0c551c2d8a856e49acd4f25794591d84ca8641d6